Test idempotence, and add some sparql tests
prepopulate twice the db and check that no other objects are inserted.

to do this we do some graphql queries.

there is a big catcha to this:
the blazegraph db is asynchronously synchronized, so in the tests, for
the moment there is an uggly std::thread::sleep(15s), to make sure the
sparql engine can find the data...