Do not search items
for the moment the `get_or_create` function used the `/wbsearchentity`
function that did text search on a label to check the existence of an
item.

We do not want this (for example we might have many items with `"bus"`
in their label).

We now then search by `topo_id` via sparql queries